prepare for package renaming
