+++ cerberus_anymal_control_1/package.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>cerberus_anymal_control_1</name>
+  <name>cerberus_anymal_b_control_1</name>
   <version>1.0.0</version>
   <description>ANYmal B controller for DARPA SubT virtual challenge.</description>
   <license>Proprietary</license>
@@ -19,5 +19,5 @@
   <depend>roscpp</depend>
   <depend>roslib</depend>
   <!-- execution dependencies -->
-  <exec_depend>cerberus_anymal_sensor_config_1</exec_depend>
+  <exec_depend>cerberus_anymal_b_sensor_config_1</exec_depend>
 </package>
